Move internal modules to _* and update imports
Move internal/private modules by renaming api/lib -> api/_lib and prefixing several auth modules with an underscore (api/auth/_*). Update all import paths (login, signup, events/register, user-storage, tests) to the new locations. Add knip.json config and adjust package.json/tests to accommodate the refactor.
