+++ package.json
@@ -73,7 +73,8 @@
     "prepare": "node scripts/setup-git-config.js",
     "setup:secret": "node scripts/setup-dev-secret.js",
     "storybook": "storybook dev -p 6006",
-    "build-storybook": "storybook build"
+    "build-storybook": "storybook build",
+    "knip": "knip"
   },
   "browserslist": {
     "production": [
@@ -106,6 +107,7 @@
     "@storybook/preset-create-react-app": "10.4.1",
     "@storybook/react-webpack5": "^10.4.1",
     "@testing-library/react": "16.3.2",
+    "@types/node": "^25.9.3",
     "@vitejs/plugin-react": "^6.0.2",
     "autoprefixer": "^10.4.21",
     "cross-env": "^10.1.0",
@@ -116,6 +118,7 @@
     "eslint-plugin-testing-library": "^7.13.5",
     "globals": "17.6.0",
     "jsdom": "^29.1.1",
+    "knip": "^6.16.1",
     "lint-staged": "^17.0.7",
     "playwright": "1.60.0",
     "postcss": "8.5.15",
