+++ src/__tests__/rate-limit-config.test.js
@@ -18,56 +18,56 @@ describe("Rate-Limit Configuration Validation", () => {
     test("should return false when KV_REST_API_URL is missing", async () => {
       delete process.env.KV_REST_API_URL;
       process.env.KV_REST_API_TOKEN = "token";
-      const { isDistributedRateLimitStorageConfigured } = await import("../../api/lib/rate-limit-config.js");
+      const { isDistributedRateLimitStorageConfigured } = await import("../../api/_lib/rate-limit-config.js");
       assert.strictEqual(isDistributedRateLimitStorageConfigured(), false);
     });
 
     test("should return false when KV_REST_API_TOKEN is missing", async () => {
       process.env.KV_REST_API_URL = "https://redis.example.com";
       delete process.env.KV_REST_API_TOKEN;
-      const { isDistributedRateLimitStorageConfigured } = await import("../../api/lib/rate-limit-config.js");
+      const { isDistributedRateLimitStorageConfigured } = await import("../../api/_lib/rate-limit-config.js");
       assert.strictEqual(isDistributedRateLimitStorageConfigured(), false);
     });
 
     test("should return false when both are missing", async () => {
       delete process.env.KV_REST_API_URL;
       delete process.env.KV_REST_API_TOKEN;
-      const { isDistributedRateLimitStorageConfigured } = await import("../../api/lib/rate-limit-config.js");
+      const { isDistributedRateLimitStorageConfigured } = await import("../../api/_lib/rate-limit-config.js");
       assert.strictEqual(isDistributedRateLimitStorageConfigured(), false);
     });
 
     test("should return false when KV_REST_API_URL is empty", async () => {
       process.env.KV_REST_API_URL = "";
       process.env.KV_REST_API_TOKEN = "token";
-      const { isDistributedRateLimitStorageConfigured } = await import("../../api/lib/rate-limit-config.js");
+      const { isDistributedRateLimitStorageConfigured } = await import("../../api/_lib/rate-limit-config.js");
       assert.strictEqual(isDistributedRateLimitStorageConfigured(), false);
     });
 
     test("should return false when KV_REST_API_TOKEN is empty", async () => {
       process.env.KV_REST_API_URL = "https://redis.example.com";
       process.env.KV_REST_API_TOKEN = "";
-      const { isDistributedRateLimitStorageConfigured } = await import("../../api/lib/rate-limit-config.js");
+      const { isDistributedRateLimitStorageConfigured } = await import("../../api/_lib/rate-limit-config.js");
       assert.strictEqual(isDistributedRateLimitStorageConfigured(), false);
     });
 
     test("should return false when KV_REST_API_URL is whitespace-only", async () => {
       process.env.KV_REST_API_URL = "   ";
       process.env.KV_REST_API_TOKEN = "token";
-      const { isDistributedRateLimitStorageConfigured } = await import("../../api/lib/rate-limit-config.js");
+      const { isDistributedRateLimitStorageConfigured } = await import("../../api/_lib/rate-limit-config.js");
       assert.strictEqual(isDistributedRateLimitStorageConfigured(), false);
     });
 
     test("should return false when KV_REST_API_TOKEN is whitespace-only", async () => {
       process.env.KV_REST_API_URL = "https://redis.example.com";
       process.env.KV_REST_API_TOKEN = "   ";
-      const { isDistributedRateLimitStorageConfigured } = await import("../../api/lib/rate-limit-config.js");
+      const { isDistributedRateLimitStorageConfigured } = await import("../../api/_lib/rate-limit-config.js");
       assert.strictEqual(isDistributedRateLimitStorageConfigured(), false);
     });
 
     test("should return true when both are present and non-empty", async () => {
       process.env.KV_REST_API_URL = "https://redis.example.com";
       process.env.KV_REST_API_TOKEN = "secret-token";
-      const { isDistributedRateLimitStorageConfigured } = await import("../../api/lib/rate-limit-config.js");
+      const { isDistributedRateLimitStorageConfigured } = await import("../../api/_lib/rate-limit-config.js");
       assert.strictEqual(isDistributedRateLimitStorageConfigured(), true);
     });
   });
@@ -77,7 +77,7 @@ describe("Rate-Limit Configuration Validation", () => {
       process.env.NODE_ENV = "production";
       delete process.env.KV_REST_API_URL;
       process.env.KV_REST_API_TOKEN = "token";
-      const { assertDistributedRateLimitStorageConfigured } = await import("../../api/lib/rate-limit-config.js");
+      const { assertDistributedRateLimitStorageConfigured } = await import("../../api/_lib/rate-limit-config.js");
       assert.throws(
         () => assertDistributedRateLimitStorageConfigured(),
         /KV_REST_API_URL and KV_REST_API_TOKEN are required in production/
@@ -88,7 +88,7 @@ describe("Rate-Limit Configuration Validation", () => {
       process.env.NODE_ENV = "production";
       process.env.KV_REST_API_URL = "https://redis.example.com";
       delete process.env.KV_REST_API_TOKEN;
-      const { assertDistributedRateLimitStorageConfigured } = await import("../../api/lib/rate-limit-config.js");
+      const { assertDistributedRateLimitStorageConfigured } = await import("../../api/_lib/rate-limit-config.js");
       assert.throws(
         () => assertDistributedRateLimitStorageConfigured(),
         /KV_REST_API_URL and KV_REST_API_TOKEN are required in production/
@@ -99,7 +99,7 @@ describe("Rate-Limit Configuration Validation", () => {
       process.env.NODE_ENV = "production";
       delete process.env.KV_REST_API_URL;
       delete process.env.KV_REST_API_TOKEN;
-      const { assertDistributedRateLimitStorageConfigured } = await import("../../api/lib/rate-limit-config.js");
+      const { assertDistributedRateLimitStorageConfigured } = await import("../../api/_lib/rate-limit-config.js");
       assert.throws(
         () => assertDistributedRateLimitStorageConfigured(),
         /KV_REST_API_URL and KV_REST_API_TOKEN are required in production/
@@ -110,7 +110,7 @@ describe("Rate-Limit Configuration Validation", () => {
       process.env.NODE_ENV = "production";
       process.env.KV_REST_API_URL = "https://redis.example.com";
       process.env.KV_REST_API_TOKEN = "secret-token";
-      const { assertDistributedRateLimitStorageConfigured } = await import("../../api/lib/rate-limit-config.js");
+      const { assertDistributedRateLimitStorageConfigured } = await import("../../api/_lib/rate-limit-config.js");
       assert.doesNotThrow(() => assertDistributedRateLimitStorageConfigured());
     });
 
@@ -118,7 +118,7 @@ describe("Rate-Limit Configuration Validation", () => {
       process.env.NODE_ENV = "development";
       delete process.env.KV_REST_API_URL;
       delete process.env.KV_REST_API_TOKEN;
-      const { assertDistributedRateLimitStorageConfigured } = await import("../../api/lib/rate-limit-config.js");
+      const { assertDistributedRateLimitStorageConfigured } = await import("../../api/_lib/rate-limit-config.js");
       assert.doesNotThrow(() => assertDistributedRateLimitStorageConfigured());
     });
 
@@ -126,7 +126,7 @@ describe("Rate-Limit Configuration Validation", () => {
       process.env.NODE_ENV = "test";
       delete process.env.KV_REST_API_URL;
       delete process.env.KV_REST_API_TOKEN;
-      const { assertDistributedRateLimitStorageConfigured } = await import("../../api/lib/rate-limit-config.js");
+      const { assertDistributedRateLimitStorageConfigured } = await import("../../api/_lib/rate-limit-config.js");
       assert.doesNotThrow(() => assertDistributedRateLimitStorageConfigured());
     });
 
@@ -134,7 +134,7 @@ describe("Rate-Limit Configuration Validation", () => {
       delete process.env.NODE_ENV;
       delete process.env.KV_REST_API_URL;
       delete process.env.KV_REST_API_TOKEN;
-      const { assertDistributedRateLimitStorageConfigured } = await import("../../api/lib/rate-limit-config.js");
+      const { assertDistributedRateLimitStorageConfigured } = await import("../../api/_lib/rate-limit-config.js");
       assert.doesNotThrow(() => assertDistributedRateLimitStorageConfigured());
     });
   });
@@ -142,25 +142,25 @@ describe("Rate-Limit Configuration Validation", () => {
   describe("isInMemoryRateLimitStorageAllowed", () => {
     test("should return true in development mode", async () => {
       process.env.NODE_ENV = "development";
-      const { isInMemoryRateLimitStorageAllowed } = await import("../../api/lib/rate-limit-config.js");
+      const { isInMemoryRateLimitStorageAllowed } = await import("../../api/_lib/rate-limit-config.js");
       assert.strictEqual(isInMemoryRateLimitStorageAllowed(), true);
     });
 
     test("should return true in test mode", async () => {
       process.env.NODE_ENV = "test";
-      const { isInMemoryRateLimitStorageAllowed } = await import("../../api/lib/rate-limit-config.js");
+      const { isInMemoryRateLimitStorageAllowed } = await import("../../api/_lib/rate-limit-config.js");
       assert.strictEqual(isInMemoryRateLimitStorageAllowed(), true);
     });
 
     test("should return false in production mode", async () => {
       process.env.NODE_ENV = "production";
-      const { isInMemoryRateLimitStorageAllowed } = await import("../../api/lib/rate-limit-config.js");
+      const { isInMemoryRateLimitStorageAllowed } = await import("../../api/_lib/rate-limit-config.js");
       assert.strictEqual(isInMemoryRateLimitStorageAllowed(), false);
     });
 
     test("should return true when NODE_ENV is not set", async () => {
       delete process.env.NODE_ENV;
-      const { isInMemoryRateLimitStorageAllowed } = await import("../../api/lib/rate-limit-config.js");
+      const { isInMemoryRateLimitStorageAllowed } = await import("../../api/_lib/rate-limit-config.js");
       assert.strictEqual(isInMemoryRateLimitStorageAllowed(), true);
     });
   });
@@ -171,7 +171,7 @@ describe("Rate-Limit Configuration Validation", () => {
       delete process.env.KV_REST_API_URL;
       delete process.env.KV_REST_API_TOKEN;
       
-      const { isInMemoryRateLimitStorageAllowed, isDistributedRateLimitStorageConfigured } = await import("../../api/lib/rate-limit-config.js");
+      const { isInMemoryRateLimitStorageAllowed, isDistributedRateLimitStorageConfigured } = await import("../../api/_lib/rate-limit-config.js");
       
       assert.strictEqual(isInMemoryRateLimitStorageAllowed(), true);
       assert.strictEqual(isDistributedRateLimitStorageConfigured(), false);
@@ -182,7 +182,7 @@ describe("Rate-Limit Configuration Validation", () => {
       delete process.env.KV_REST_API_URL;
       delete process.env.KV_REST_API_TOKEN;
       
-      const { isInMemoryRateLimitStorageAllowed, isDistributedRateLimitStorageConfigured } = await import("../../api/lib/rate-limit-config.js");
+      const { isInMemoryRateLimitStorageAllowed, isDistributedRateLimitStorageConfigured } = await import("../../api/_lib/rate-limit-config.js");
       
       assert.strictEqual(isInMemoryRateLimitStorageAllowed(), true);
       assert.strictEqual(isDistributedRateLimitStorageConfigured(), false);
