+++ src/__tests__/rate-limit-storage.test.js
@@ -15,27 +15,27 @@ describe("Rate-Limit Storage (In-Memory)", () => {
   afterEach(async () => {
     process.env = originalEnv;
     process.env.NODE_ENV = originalNodeEnv;
-    const { clearAll } = await import("../../api/lib/rate-limit-storage.js");
+    const { clearAll } = await import("../../api/_lib/rate-limit-storage.js");
     await clearAll();
   });
 
   describe("incrementWithExpiration", () => {
     test("should increment counter for new key", async () => {
-      const { incrementWithExpiration } = await import("../../api/lib/rate-limit-storage.js");
+      const { incrementWithExpiration } = await import("../../api/_lib/rate-limit-storage.js");
       const result = await incrementWithExpiration("test-key", 60000);
       assert.strictEqual(result.count, 1);
       assert.ok(result.ttl > 0);
     });
 
     test("should increment counter for existing key within window", async () => {
-      const { incrementWithExpiration } = await import("../../api/lib/rate-limit-storage.js");
+      const { incrementWithExpiration } = await import("../../api/_lib/rate-limit-storage.js");
       await incrementWithExpiration("test-key", 60000);
       const result = await incrementWithExpiration("test-key", 60000);
       assert.strictEqual(result.count, 2);
     });
 
     test("should reset counter after window expires", async () => {
-      const { incrementWithExpiration } = await import("../../api/lib/rate-limit-storage.js");
+      const { incrementWithExpiration } = await import("../../api/_lib/rate-limit-storage.js");
       await incrementWithExpiration("test-key", 100); // 100ms window
       await new Promise(resolve => setTimeout(resolve, 150)); // Wait for expiration
       const result = await incrementWithExpiration("test-key", 100);
@@ -43,7 +43,7 @@ describe("Rate-Limit Storage (In-Memory)", () => {
     });
 
     test("should handle multiple keys independently", async () => {
-      const { incrementWithExpiration } = await import("../../api/lib/rate-limit-storage.js");
+      const { incrementWithExpiration } = await import("../../api/_lib/rate-limit-storage.js");
       const result1 = await incrementWithExpiration("key1", 60000);
       const result2 = await incrementWithExpiration("key2", 60000);
       assert.strictEqual(result1.count, 1);
@@ -51,7 +51,7 @@ describe("Rate-Limit Storage (In-Memory)", () => {
     });
 
     test("should return remaining TTL", async () => {
-      const { incrementWithExpiration } = await import("../../api/lib/rate-limit-storage.js");
+      const { incrementWithExpiration } = await import("../../api/_lib/rate-limit-storage.js");
       const windowMs = 60000;
       const result = await incrementWithExpiration("test-key", windowMs);
       assert.ok(result.ttl > 0);
@@ -61,7 +61,7 @@ describe("Rate-Limit Storage (In-Memory)", () => {
 
   describe("resetKey", () => {
     test("should reset a specific key", async () => {
-      const { incrementWithExpiration, resetKey } = await import("../../api/lib/rate-limit-storage.js");
+      const { incrementWithExpiration, resetKey } = await import("../../api/_lib/rate-limit-storage.js");
       await incrementWithExpiration("test-key", 60000);
       await incrementWithExpiration("test-key", 60000);
       await resetKey("test-key");
@@ -70,7 +70,7 @@ describe("Rate-Limit Storage (In-Memory)", () => {
     });
 
     test("should handle resetting non-existent key", async () => {
-      const { resetKey, incrementWithExpiration } = await import("../../api/lib/rate-limit-storage.js");
+      const { resetKey, incrementWithExpiration } = await import("../../api/_lib/rate-limit-storage.js");
       await assert.doesNotReject(resetKey("non-existent-key"));
       const result = await incrementWithExpiration("non-existent-key", 60000);
       assert.strictEqual(result.count, 1);
@@ -79,7 +79,7 @@ describe("Rate-Limit Storage (In-Memory)", () => {
 
   describe("clearAll", () => {
     test("should clear all keys", async () => {
-      const { incrementWithExpiration, clearAll } = await import("../../api/lib/rate-limit-storage.js");
+      const { incrementWithExpiration, clearAll } = await import("../../api/_lib/rate-limit-storage.js");
       await incrementWithExpiration("key1", 60000);
       await incrementWithExpiration("key2", 60000);
       await incrementWithExpiration("key3", 60000);
@@ -95,7 +95,7 @@ describe("Rate-Limit Storage (In-Memory)", () => {
 
   describe("Concurrent requests", () => {
     test("should handle concurrent increments correctly", async () => {
-      const { incrementWithExpiration, resetKey } = await import("../../api/lib/rate-limit-storage.js");
+      const { incrementWithExpiration, resetKey } = await import("../../api/_lib/rate-limit-storage.js");
       const key = "concurrent-key";
       
       // Make 10 concurrent requests
@@ -137,7 +137,7 @@ describe("Rate-Limit Storage (Production Fail-Closed)", () => {
 
   test("should throw error in production without distributed storage", async () => {
     process.env.NODE_ENV = "production";
-    const { incrementWithExpiration } = await import("../../api/lib/rate-limit-storage.js");
+    const { incrementWithExpiration } = await import("../../api/_lib/rate-limit-storage.js");
     await assert.rejects(
       incrementWithExpiration("test-key", 60000),
       /Distributed rate-limit storage is required in production/
@@ -146,14 +146,14 @@ describe("Rate-Limit Storage (Production Fail-Closed)", () => {
 
   test("should allow in-memory storage in development without distributed storage", async () => {
     process.env.NODE_ENV = "development";
-    const { incrementWithExpiration } = await import("../../api/lib/rate-limit-storage.js");
+    const { incrementWithExpiration } = await import("../../api/_lib/rate-limit-storage.js");
     const result = await incrementWithExpiration("test-key", 60000);
     assert.strictEqual(result.count, 1);
   });
 
   test("should allow in-memory storage in test without distributed storage", async () => {
     process.env.NODE_ENV = "test";
-    const { incrementWithExpiration, clearAll } = await import("../../api/lib/rate-limit-storage.js");
+    const { incrementWithExpiration, clearAll } = await import("../../api/_lib/rate-limit-storage.js");
     await clearAll(); // Clear any existing state
     const result = await incrementWithExpiration("test-key", 60000);
     assert.strictEqual(result.count, 1);
